Updated for itch.io release
updated scoreboard

+++ Assets/Scripts/ScoreHolder.cs
@@ -6,9 +6,9 @@ public class ScoreHolder : MonoBehaviour
 {
 
     public string playerName;
-    public int score;
+    public float score;
 
-    public ScoreHolder(string _name, int _score)
+    public ScoreHolder(string _name, float _score)
     {
         playerName = _name;
         score = _score;
